frontend/login: Cookie login
