+++ package.json
@@ -18,6 +18,7 @@
     "@babel/preset-typescript": "^7.10.4",
     "@types/fast-text-encoding": "^1.0.1",
     "@types/jest": "^26.0.14",
+    "@types/js-cookie": "^2.2.6",
     "@types/react": "^16.9.52",
     "@types/react-dom": "^16.9.8",
     "@types/react-router": "^5.1.8",
@@ -49,6 +50,7 @@
     "axios": "^0.20.0",
     "bulma": "^0.9.1",
     "fast-text-encoding": "^1.0.3",
+    "js-cookie": "^2.2.1",
     "jsbi": "^3.1.4",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
